Make ESM Instance ID Configurable
Feature request to allow practitioners to set instance ids in order for ESM
instances to terminate and reregister with Consul with the same id. Ungraceful
terminations currently take 30 minutes to be reaped. Without this feature, when
a new instance starts up, it has a new id and leads to multiple ESM instances
rather than 'replacing' the terminated one.

- Add instance_id in configuration file
- Replace existing `id` used for testing with instance_id
- Add a check to ensure instance_id is unique since we are no longer generating
a unique id on behalf of the practitioner