Add “operator=read” to ACL documentation
ACL for operator:read is necessary for Consul-ESM v0.4.0 due to a new feature that check’s ESM and Consul’s version compatibility on start up: https://github.com/hashicorp/consul-esm/pull/62. This feature makes a request to /operator/autopilot, which needs to be added to the list of necessary ACLs.