change operator call to catalog call
Swap out Operator.AutopilotServerHealth with Catalog.Service for
determining the version of the consul servers in order to stop using the
Operator based call as it requires "operator: read" ACL permissions. We
don't want "operator: read" as that means ESM can read all Namespaces
and we need to allow for limiting Namespace access via ACLs.