fix bug with deregister->register
Bug desc: Deregister a node then re-registering it quickly would lead to
a node without coordiates, etc. for a time as there is a timed based
status entry that needed to expire before the re-registration would be
complete with coordinates and all.

This change clears out those saved status along with the making sure the
inflight ping tracking is also cleared. So that the re-registered node
can be processed normally.