don't depend on goroutine to avoid deadlock.
Code calls a function that has a lock that is held around that call. It
didn't deadlock as the function call was calling it as a goroutine, so
it would kick it off and then unlock which would let the goroutine then
get the lock. Found this when I temporarily removed to goroutine to
force sequential exectution for debugging. Didn't like this so I cleaned
things up such that the code doesn't rely on the goroutine to avoid the
deadlock.