update to use Consul's new StatusHandler, et. al.
StatusHandler is an internal public struct used to track service status.
This replaces the older Notify interface, subsuming it as a field.

The Notify interface itself (used on a field of the StatusHandler) also
has added a new method, ServiceExists, which was added as part of the
Consul's Alias feature and which we don't support so use add a stubbed
out method.

The IDs used to track the checks have also changed from a string to
a struct. The struct contains the string and some enterprise meta data.
For now we just ignore the enterprise portion internally.