fix issue with timeouts on startup
During initial startup ESM would spam the KV() endpoint for data without
any sort of retry delay. This was causing the limited CPU on the test
instances to get used up.

Fixed this by adding a sleep for the 'retryTime' as is done in other
places (IMO this case was just missed).