fix deadlock in check runner
CheckRunner had a race where it would deadlock if the right timing of
Stop() and an internal run of the check. They'd lock on the
checkrunner's Mutex.

This replaces the 2 maps used in Stop from being protected by the mutext
to sync.Map instances. They handle the mutex all internally and don't
need to use the CR one so they no longer trigger the deadlock on Stop.

The CR really needs to be reworked a bit to minimize/eliminate the
internal state maps and the need for the lock, but that is a large
refactor while this fixes it with minimal changes.